backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase>=2.0.0
httpx>=0.25.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
slowapi==0.1.9
reportlab==4.0.7
qrcode[pil]==7.4.2
Pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
numpy>=1.24.0
opencv-python>=4.7.0
librosa>=0.10.0
soundfile>=0.12.0
